requirements.txt
python-dotenv
PyPDF2
azure-search-documents
openai
pypdf